introduced _core and _core_class as well as modified arguments to get_unformatted_data
