+++ sami2py/tests/test_core.py
