+++ sami2py/utils.py
@@ -7,12 +7,14 @@
 
 Functions
 -------------------------------------------------------------------------------
-
 generate_path(tag, lon, year, day)
     Generates path to archived model runs based on input paramters.
 
 set_archive_dir(path=None, store=None)
     Allows user to specify the location where the model outputs will be stored
+
+get_unformatted_data(dat_dir, var_name, nz, nf, ni, nt, reshape=False)
+    routine to interpret unformatted binary files created by the SAMI2 model
 -------------------------------------------------------------------------------
 
 Moduleauthor
@@ -93,3 +95,40 @@ def set_archive_dir(path=None, store=True):
         sami2py.archive_dir = path
     else:
         raise ValueError('Path does not lead to a valid directory.')
+
+def get_unformatted_data(dat_dir, var_name, reshape=False, dim0=0, dim1=0):
+    """Routine to interpret unformatted binary files created by the SAMI2 model
+
+    Parameters
+    -----------
+    data_dir : (str)
+        directory where the SAMI2 data is stored
+    var_name : (str)
+        name of unformatted data variable to be loaded
+    nz : (int)
+        number of mesh points along the geomagnetic field line
+    nf : (int)
+        number of mesh points transverse to the geomagnetic field line i.e.
+        number of field lines
+    ni : (int)
+        number of ion species
+    nt : (int)
+        number of time steps
+    reshape : (bool)
+        if true the data is reshaped by the mesh geometry
+
+    Returns
+    -----------
+    float_data : (numpy.ndarray)
+        unformatted data organized into a numpy array for handling in python
+    """
+    import numpy as np
+    from os import path
+    binary_file = open(path.join(dat_dir, var_name + 'u.dat'), 'rb')
+    float_data = np.fromfile(binary_file, dtype='float32')
+    binary_file.close()
+
+    if reshape:
+        float_data.shape = (dim0, dim1)
+        return float_data[1:-1, :]
+    return float_data[1:-1]
